refactors Projects component into Project sub-component and two parent components: ProjectsPage and ProjectsHighlight
