+++ bin/seed.js
@@ -41,61 +41,49 @@ const data = {
     {
       name: 'd3TopSearchedArtists1',
       imagePath: 'assets/projects/d3_top_searched_artists/screenshot1.jpg',
-      titlePath: '',
       category: 'project',
       projectName: 'D3.js Visualization://Top Searched Visual Artists',
       description: 'Data visualizations using D3.js of the top 20 most searched visual artists on artnet.com in Sept 2017. Visualizations include: tree map, bubble chart, and plain list.',
       technologies: 'JavaScript - React, D3; Node, Express; HTML/SCSS',
       linkProject: 'https://top-searched-artists-09-2017.herokuapp.com',
-      linkDemo: '',
       linkGitHub: 'https://github.com/jrleja0/D3-Visualization-Top-Searched-Artists',
     },
     {
       name: 'easyPressReleaseMaker1',
       imagePath: 'assets/projects/easy_press_release_maker/screenshot1.jpg',
-      titlePath: '',
       category: 'project',
       projectName: 'Easy Press Release Maker',
       description: 'A prototype fullstack RESTful CRUD app that creates press releases quickly and easily.',
       technologies: 'JavaScript - React, Redux; Node, Express, and PostgreSQL/Sequelize; HTML/SCSS.',
-      linkProject: '',
-      linkDemo: '',
       linkGitHub: 'https://github.com/jrleja0/Easy-Press-Release-Maker',
     },
     {
       name: 'quickLookChromeExt1',
       imagePath: 'assets/projects/quick_look_chrome_ext/screenshot1.jpg',
-      titlePath: '',
       category: 'project',
       projectName: '"Quick Look" Google Search//Chrome Extension',
       description: 'A simple Chrome extension that searches Google for top titles related to your search.',
       technologies: 'JavaScript - React, Axios, Cheerio, Webpack; HTML/CSS',
-      linkProject: '',
-      linkDemo: '',
       linkGitHub: 'https://github.com/jrleja0/quick-look-google-search-chrome-extension',
     },
     {
       name: 'tileSwappingPuzzleGame1',
       imagePath: 'assets/projects/tile_swapping_puzzle_game/screenshot1.jpg',
-      titlePath: '',
       category: 'project',
       projectName: 'Tile Swapping Puzzle Game',
       description: 'HTML5 Canvas & JavaScript puzzle game. Click two tiles to swap them. Swap tiles until the image is unscrambled.',
       technologies: 'JavaScript, HTML5 Canvas; HTML/CSS',
       linkProject: 'https://jrleja0.github.io/tile-swapping-game-jrleja0',
-      linkDemo: '',
       linkGitHub: 'https://github.com/jrleja0/tile-swapping-game-jrleja0',
     },
     {
       name: 'movingTilesAnimation1',
       imagePath: 'assets/projects/moving_tiles_animation/screenshot1.jpg',
-      titlePath: '',
       category: 'project',
       projectName: 'Moving Tiles Animation',
       description: 'HTML5 Canvas & JavaScript animation.',
       technologies: 'JavaScript, HTML5 Canvas; HTML/CSS',
       linkProject: 'https://jrleja0.github.io/Tile-Abstraction-v2',
-      linkDemo: '',
       linkGitHub: 'https://github.com/jrleja0/Tile-Abstraction-v2',
     },
     {
