+++ client/Routes.js
@@ -3,7 +3,7 @@ import {connect} from 'react-redux';
 import {Redirect, Route, Switch, withRouter} from 'react-router-dom';
 import PropTypes from 'prop-types';
 import {TransitionGroup, CSSTransition} from 'react-transition-group';
-import {ArtShowcase, Contact, Main, Projects, Slideshow, Welcome} from './components';
+import {ArtShowcase, Contact, Main, ProjectsHighlight, ProjectsPage, Slideshow, Welcome} from './components';
 import { fetchImages } from './store';
 
 const pathBackgroundColors = {
@@ -17,7 +17,7 @@ const Home = () => (
   <div className="green-fixed-background">
     <Slideshow />
     <Welcome />
-    <Projects />
+    <ProjectsHighlight />
     <div className="red-fixed-small-background">
       <Contact />
     </div>
@@ -50,7 +50,7 @@ class Routes extends Component {
             <Main pathName={location.pathname} pathBackgroundColors={pathBackgroundColors}>
               <Switch location={location}>
                 <Route path="/home" component={Home} />
-                <Route path="/projects" component={Projects} />
+                <Route path="/projects" component={ProjectsPage} />
                 <Route path="/art" component={ArtShowcase} />
                 <Route path="/contact" component={Contact} />
                 <Redirect to="/home" />
