+++ client/components/index.js
@@ -2,7 +2,9 @@ export {default as Main} from './Main';
 export {default as MainLogo} from './MainLogo';
 export {default as ArtShowcase} from './ArtShowcase';
 export {default as Contact} from './Contact';
-export {default as Projects} from './Projects';
+export {default as Project} from './Project';
+export {default as ProjectsHighlight} from './ProjectsHighlight';
+export {default as ProjectsPage} from './ProjectsPage';
 export {default as Slideshow} from './Slideshow';
 export {default as Spinner} from './Spinner';
 export {default as Welcome} from './Welcome';
